+++ src/gfx/online_renderer.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "model_loading/model_internal.h"
-#include "gfx/gfx.h"
+#include "gfx_api/gfx.h"
 
 struct offline_to_online_vertex_t {
   vec2 position;
